Split desktop controller routing extensions
Keep QuillCodeDesktopController focused on bootstrap state and refresh while moving UI action families into small controller extensions. Update parity gates to audit the full controller family and prevent the core file from growing back into a routing catch-all.
