+++ Sources/quill-code-desktop/QuillCodeDesktopController+Terminal.swift
@@ -1,0 +1,48 @@
+import QuillCodeApp
+
+@MainActor
+extension QuillCodeDesktopController {
+    func runTerminalCommand() {
+        terminalCoordinator.runCommand(
+            draft: &terminalDraft,
+            model: model,
+            fallbackWorkspaceRoot: workspaceRoot,
+            tasks: tasks,
+            refresh: { [weak self] in self?.refresh() }
+        )
+    }
+
+    func recallPreviousTerminalCommand() {
+        terminalCoordinator.recallPreviousCommand(
+            draft: &terminalDraft,
+            model: model,
+            tasks: tasks,
+            refresh: { [weak self] in self?.refresh() }
+        )
+    }
+
+    func recallNextTerminalCommand() {
+        terminalCoordinator.recallNextCommand(
+            draft: &terminalDraft,
+            model: model,
+            tasks: tasks,
+            refresh: { [weak self] in self?.refresh() }
+        )
+    }
+
+    func resizeTerminal(_ windowSize: TerminalWindowSize) {
+        terminalCoordinator.resizeTerminal(windowSize, model: model)
+    }
+
+    func suspendTerminal() {
+        terminalCoordinator.suspendTerminal(model: model)
+        // Rebuild the published surface so the pane re-renders with the new isSuspended state
+        // (Resume replaces Suspend). isSuspended is a visible surface field, unlike resize.
+        refresh()
+    }
+
+    func resumeTerminal() {
+        terminalCoordinator.resumeTerminal(model: model)
+        refresh()
+    }
+}
