+++ Sources/quill-code-desktop/QuillCodeDesktopController.swift
@@ -16,29 +16,29 @@ final class QuillCodeDesktopController: ObservableObject {
     @Published var isProjectImporterPresented = false
     @Published var copiedTranscriptItemID: String?
 
-    private let model: QuillCodeWorkspaceModel
-    private let bootstrap: QuillCodeWorkspaceBootstrap
-    private let computerUseCoordinator: QuillCodeDesktopComputerUseCoordinator
-    private let activeWorkCoordinator: QuillCodeDesktopActiveWorkCoordinator
-    private let browserCoordinator: QuillCodeDesktopBrowserCoordinator
-    private let automationCoordinator: QuillCodeDesktopAutomationCoordinator
-    private let automationNotifier: any QuillCodeAutomationNotifying
-    private let workspaceRoot: URL
-    private let navigationCoordinator: QuillCodeDesktopNavigationCoordinator
-    private let commandCoordinator: QuillCodeDesktopCommandCoordinator
-    private let signInCoordinator: QuillCodeDesktopSignInCoordinator
-    private let settingsCoordinator: QuillCodeDesktopSettingsCoordinator
-    private let modelCatalogRefreshCoordinator: QuillCodeDesktopModelCatalogRefreshCoordinator
-    private let composerCoordinator: QuillCodeDesktopComposerCoordinator
-    private let copyCoordinator: QuillCodeDesktopCopyCoordinator
-    private let projectImportCoordinator: QuillCodeDesktopProjectImportCoordinator
-    private let modelStateCoordinator: QuillCodeDesktopModelStateCoordinator
-    private let paneCoordinator: QuillCodeDesktopPaneCoordinator
-    private let workspaceActionCoordinator: QuillCodeDesktopWorkspaceActionCoordinator
-    private let terminalCoordinator: QuillCodeDesktopTerminalCoordinator
-    private let transcriptExportCoordinator: QuillCodeDesktopTranscriptExportCoordinator
-    private let worktreeCoordinator: QuillCodeDesktopWorktreeCoordinator
-    private let tasks = QuillCodeDesktopTaskCoordinator()
+    let model: QuillCodeWorkspaceModel
+    let bootstrap: QuillCodeWorkspaceBootstrap
+    let computerUseCoordinator: QuillCodeDesktopComputerUseCoordinator
+    let activeWorkCoordinator: QuillCodeDesktopActiveWorkCoordinator
+    let browserCoordinator: QuillCodeDesktopBrowserCoordinator
+    let automationCoordinator: QuillCodeDesktopAutomationCoordinator
+    let automationNotifier: any QuillCodeAutomationNotifying
+    let workspaceRoot: URL
+    let navigationCoordinator: QuillCodeDesktopNavigationCoordinator
+    let commandCoordinator: QuillCodeDesktopCommandCoordinator
+    let signInCoordinator: QuillCodeDesktopSignInCoordinator
+    let settingsCoordinator: QuillCodeDesktopSettingsCoordinator
+    let modelCatalogRefreshCoordinator: QuillCodeDesktopModelCatalogRefreshCoordinator
+    let composerCoordinator: QuillCodeDesktopComposerCoordinator
+    let copyCoordinator: QuillCodeDesktopCopyCoordinator
+    let projectImportCoordinator: QuillCodeDesktopProjectImportCoordinator
+    let modelStateCoordinator: QuillCodeDesktopModelStateCoordinator
+    let paneCoordinator: QuillCodeDesktopPaneCoordinator
+    let workspaceActionCoordinator: QuillCodeDesktopWorkspaceActionCoordinator
+    let terminalCoordinator: QuillCodeDesktopTerminalCoordinator
+    let transcriptExportCoordinator: QuillCodeDesktopTranscriptExportCoordinator
+    let worktreeCoordinator: QuillCodeDesktopWorktreeCoordinator
+    let tasks = QuillCodeDesktopTaskCoordinator()
     // Retained here because UNUserNotificationCenter.delegate is weak; nil until the window installs it.
     private var approvalNotificationDelegate: QuillCodeApprovalNotificationDelegate?
 
@@ -159,95 +159,6 @@ final class QuillCodeDesktopController: ObservableObject {
         }
     }
 
-    func newChat() {
-        // Stash the live draft to the outgoing thread before opening a new chat.
-        model.setDraft(draft)
-        navigationCoordinator.newChat(model: model)
-        modelStateCoordinator.syncComposerDraft(from: model, draft: &draft)
-        refresh()
-    }
-
-    func selectThread(_ id: UUID) {
-        // Persist the live in-progress draft to the outgoing thread before switching
-        // so the model can stash it and restore the incoming thread's draft.
-        model.setDraft(draft)
-        navigationCoordinator.selectThread(id, model: model)
-        // Force-sync the restored draft even mid-send (the busy gate would otherwise
-        // keep the old thread's text visible under the newly selected thread).
-        modelStateCoordinator.syncComposerDraft(from: model, draft: &draft)
-        refresh()
-    }
-
-    func runThreadAction(_ mutation: WorkspaceThreadRowMutation) {
-        // Stash the live draft before a row action (duplicate/archive/delete) that
-        // may change the selected thread, then force-sync the restored draft.
-        model.setDraft(draft)
-        navigationCoordinator.runThreadAction(mutation, model: model)
-        modelStateCoordinator.syncComposerDraft(from: model, draft: &draft)
-        refresh()
-    }
-
-    func renameThread(_ id: UUID, title: String) {
-        _ = navigationCoordinator.renameThread(id, title: title, model: model)
-        refresh()
-    }
-
-    func saveSidebarSavedSearch(title: String, query: String) {
-        _ = model.saveSidebarSavedSearch(title: title, query: query)
-        refresh()
-    }
-
-    func selectProject(_ id: UUID?) {
-        navigationCoordinator.selectProject(id, model: model)
-        refresh()
-    }
-
-    func runProjectAction(_ mutation: WorkspaceProjectRowMutation) {
-        navigationCoordinator.runProjectAction(mutation, model: model)
-        refresh()
-    }
-
-    func renameProject(_ id: UUID, name: String) {
-        _ = navigationCoordinator.renameProject(id, name: name, model: model)
-        refresh()
-    }
-
-    func requestAddProject() {
-        isProjectImporterPresented = true
-    }
-
-    func handleProjectImport(_ result: Result<[URL], Error>) {
-        guard let selection = projectImportCoordinator.selectedProject(from: result) else {
-            return
-        }
-        addProject(selection.url)
-    }
-
-    func addProject(_ url: URL) {
-        navigationCoordinator.addProject(url, model: model)
-        refresh()
-    }
-
-    func setMode(_ mode: AgentMode) {
-        settingsCoordinator.setMode(mode, on: model)
-        refresh()
-    }
-
-    func setModel(_ modelID: String) {
-        settingsCoordinator.setModel(modelID, on: model)
-        refresh()
-    }
-
-    func toggleModelFavorite(_ modelID: String) {
-        settingsCoordinator.toggleModelFavorite(modelID, on: model)
-        refresh()
-    }
-
-    func refreshModelCatalog() async {
-        await settingsCoordinator.refreshModelCatalog(on: model)
-        refresh()
-    }
-
     private func scheduleModelCatalogRefreshIfNeeded() {
         tasks.startIfIdle(.modelCatalogRefresh) { [weak self] in
             guard let self else { return }
@@ -258,327 +169,7 @@ final class QuillCodeDesktopController: ObservableObject {
         }
     }
 
-    func saveSettings(_ update: WorkspaceSettingsUpdate) {
-        settingsCoordinator.saveSettings(
-            update,
-            to: model,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func startTrustedRouterSignIn() {
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            await signInCoordinator.completeSignInAndApply(
-                to: model,
-                settingsCoordinator: settingsCoordinator,
-                refresh: { [weak self] in self?.refresh() }
-            )
-        }
-    }
-
-    func send() {
-        composerCoordinator.send(
-            draft: &draft,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func retryLastTurn() {
-        composerCoordinator.retryLastTurn(
-            draft: &draft,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func runCommand(_ command: WorkspaceCommandSurface) {
-        guard let action = QuillCodeDesktopCommandPlanner.action(for: command) else { return }
-        commandCoordinator.run(action, performer: self)
-    }
-
-    func toggleTerminal() {
-        paneCoordinator.toggleTerminal(on: model)
-        refresh()
-    }
-
-    func toggleBrowser() {
-        paneCoordinator.toggleBrowser(on: model)
-        refresh()
-    }
-
-    func toggleExtensions() {
-        paneCoordinator.toggleExtensions(on: model)
-        refresh()
-    }
-
-    func toggleMemories() {
-        paneCoordinator.toggleMemories(on: model)
-        refresh()
-    }
-
-    func toggleActivity() {
-        paneCoordinator.toggleActivity(on: model)
-        refresh()
-    }
-
-    func toggleAutomations() {
-        paneCoordinator.toggleAutomations(on: model)
-        refresh()
-    }
-
-    func openBrowserPreview() {
-        browserCoordinator.openPreview(
-            model: model,
-            addressDraft: browserAddressDraft,
-            workspaceRoot: workspaceRoot,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func openBrowserSession() {
-        browserCoordinator.openSession(
-            model: model,
-            addressDraft: browserAddressDraft,
-            workspaceRoot: workspaceRoot,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func addBrowserComment(_ comment: String) {
-        paneCoordinator.addBrowserComment(comment, to: model)
-        refresh()
-    }
-
-    func runToolCardAction(_ action: ToolCardActionSurface) {
-        workspaceActionCoordinator.runToolCardAction(
-            action,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-        refresh()
-    }
-
-    func runTurnRevert(_ turnMessageID: UUID) {
-        workspaceActionCoordinator.runTurnRevert(
-            turnMessageID: turnMessageID,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot
-        )
-        refresh()
-    }
-
-    func runTerminalCommand() {
-        terminalCoordinator.runCommand(
-            draft: &terminalDraft,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func recallPreviousTerminalCommand() {
-        terminalCoordinator.recallPreviousCommand(
-            draft: &terminalDraft,
-            model: model,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func recallNextTerminalCommand() {
-        terminalCoordinator.recallNextCommand(
-            draft: &terminalDraft,
-            model: model,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func resizeTerminal(_ windowSize: TerminalWindowSize) {
-        terminalCoordinator.resizeTerminal(windowSize, model: model)
-    }
-
-    func suspendTerminal() {
-        terminalCoordinator.suspendTerminal(model: model)
-        // Rebuild the published surface so the pane re-renders with the new isSuspended state
-        // (Resume replaces Suspend). isSuspended is a visible surface field, unlike resize.
-        refresh()
-    }
-
-    func resumeTerminal() {
-        terminalCoordinator.resumeTerminal(model: model)
-        refresh()
-    }
-
-    func runReviewAction(_ action: WorkspaceReviewActionSurface) {
-        workspaceActionCoordinator.runReviewAction(action, model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func runPullRequestReviewThreadAction(_ action: WorkspacePullRequestReviewThreadActionSurface) {
-        workspaceActionCoordinator.runPullRequestReviewThreadAction(
-            action,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot
-        )
-        refresh()
-    }
-
-    func runPullRequestReviewThreadReply(_ request: WorkspacePullRequestReviewThreadReplyRequest) {
-        workspaceActionCoordinator.runPullRequestReviewThreadReply(
-            request,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot
-        )
-        refresh()
-    }
-
-    func updatePullRequestReviewDraft(_ draft: WorkspacePullRequestReviewDraftSurface) {
-        workspaceActionCoordinator.updatePullRequestReviewDraft(draft, model: model)
-        refresh()
-    }
-
-    func cancelPullRequestReviewDraft() {
-        workspaceActionCoordinator.cancelPullRequestReviewDraft(model: model)
-        refresh()
-    }
-
-    func submitPullRequestReviewDraft() {
-        workspaceActionCoordinator.submitPullRequestReviewDraft(model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func usePullRequestReviewThreadReplyDraft(_ draft: String) {
-        workspaceActionCoordinator.useComposerDraft(draft, model: model)
-        refresh()
-    }
-
-    func addReviewComment(
-        path: String,
-        lineNumber: Int?,
-        endLineNumber: Int?,
-        lineKind: WorkspaceReviewLineKind?,
-        text: String
-    ) {
-        workspaceActionCoordinator.addReviewComment(
-            path: path,
-            lineNumber: lineNumber,
-            endLineNumber: endLineNumber,
-            lineKind: lineKind,
-            text: text,
-            model: model
-        )
-        refresh()
-    }
-
-    func createWorktree(_ request: WorkspaceWorktreeCreateRequest) {
-        worktreeCoordinator.createWorktree(request, model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func worktreeChoiceLoad() async -> WorkspaceWorktreeChoiceLoad {
-        await worktreeCoordinator.worktreeChoiceLoad(model: model, fallbackWorkspaceRoot: workspaceRoot)
-    }
-
-    func worktreePrunePreview() async -> WorkspaceWorktreePrunePreview {
-        await worktreeCoordinator.worktreePrunePreview(model: model, fallbackWorkspaceRoot: workspaceRoot)
-    }
-
-    func openWorktree(_ request: WorkspaceWorktreeOpenRequest) {
-        worktreeCoordinator.openWorktree(request, model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func removeWorktree(_ request: WorkspaceWorktreeRemoveRequest) {
-        worktreeCoordinator.removeWorktree(request, model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func pruneWorktrees(_ request: WorkspaceWorktreePruneRequest) {
-        worktreeCoordinator.pruneWorktrees(request, model: model, fallbackWorkspaceRoot: workspaceRoot)
-        refresh()
-    }
-
-    func copyTranscriptItem(id: String, text: String) {
-        guard let feedback = copyCoordinator.copyTranscriptItem(id: id, text: text) else { return }
-        copiedTranscriptItemID = feedback.copiedTranscriptItemID
-        Task { [weak self] in
-            try? await Task.sleep(nanoseconds: feedback.clearAfterNanoseconds)
-            await MainActor.run {
-                if self?.copiedTranscriptItemID == id {
-                    self?.copiedTranscriptItemID = nil
-                }
-            }
-        }
-    }
-
-    func exportCurrentConversationMarkdown() {
-        guard let markdown = TranscriptMarkdownExporter.exportableMarkdown(for: surface.transcript) else {
-            return
-        }
-        exportConversationMarkdown(title: surface.topBar.primaryTitle, markdown: markdown)
-    }
-
-    func exportConversationMarkdown(title: String, markdown: String) {
-        do {
-            _ = try transcriptExportCoordinator.exportConversation(title: title, markdown: markdown)
-        } catch {
-            model.setAgentStatus(
-                TopBarAgentStatusLabel.failed,
-                lastError: "Could not export conversation: \(error.localizedDescription)"
-            )
-            refresh()
-        }
-    }
-
-    func setMessageFeedback(messageID: UUID, value: MessageFeedbackValue) {
-        guard workspaceActionCoordinator.setMessageFeedback(messageID: messageID, value: value, model: model) else {
-            return
-        }
-        refresh()
-    }
-
-    func openCommandPalette() {
-        isCommandPalettePresented = true
-    }
-
-    func openKeyboardShortcuts() {
-        isKeyboardShortcutsPresented = true
-    }
-
-    func openSettings() {
-        isSettingsPresented = true
-    }
-
-    func stopAll() {
-        activeWorkCoordinator.stopAll(
-            draft: &draft,
-            model: model,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    func disconnectAll() {
-        activeWorkCoordinator.disconnectAll(
-            draft: &draft,
-            model: model,
-            tasks: tasks,
-            refresh: { [weak self] in self?.refresh() }
-        )
-    }
-
-    private func refresh() {
+    func refresh() {
         computerUseCoordinator.refreshStatus(on: model)
         modelStateCoordinator.refreshState(
             from: model,
@@ -588,32 +179,5 @@ final class QuillCodeDesktopController: ObservableObject {
             browserAddressDraft: &browserAddressDraft,
             isComposerTaskRunning: tasks.isRunning(.send)
         )
-    }
-
-    func openComputerUseSystemSettings(_ destination: MacSystemSettingsOpener.Destination) {
-        computerUseCoordinator.openSystemSettings(destination, model: model)
-        refresh()
-    }
-}
-
-extension QuillCodeDesktopController: QuillCodeDesktopCommandPerforming {
-    func refreshComputerUseStatus() {
-        refresh()
-    }
-
-    func runWorkspaceCommand(_ commandID: String) {
-        // Push the live draft so thread-changing commands (new chat / duplicate /
-        // fork / compact) stash it instead of reading a stale model draft.
-        model.setDraft(draft)
-        guard workspaceActionCoordinator.runWorkspaceCommand(
-            commandID,
-            model: model,
-            fallbackWorkspaceRoot: workspaceRoot
-        ) else {
-            return
-        }
-        modelStateCoordinator.syncComposerDraft(from: model, draft: &draft)
-        browserCoordinator.syncOpenSession(model: model)
-        refresh()
     }
 }
